requirements.txt
discord.py
yt-dlp
flask
python-dotenv
